Refactorized.
